finalized delete wallet

+++ app/Containers/Wallet/Tasks/DeleteWalletTask.php
@@ -3,9 +3,12 @@
 namespace App\Containers\Wallet\Tasks;
 
 use App\Containers\Wallet\Data\Repositories\WalletRepository;
+use App\Containers\Wallet\Exceptions\CouldNotDeleteWalletWithBalanceException;
+use App\Containers\Wallet\Exceptions\WalletNotFoundException;
 use App\Ship\Exceptions\DeleteResourceFailedException;
 use App\Ship\Parents\Tasks\Task;
 use Exception;
+use Tartan\Log\Facades\XLog;
 
 class DeleteWalletTask extends Task
 {
@@ -19,10 +22,23 @@ class DeleteWalletTask extends Task
 
     public function run($id)
     {
+        $wallet = $this->repository->find($id);
+
+        if (empty($wallet)) {
+            throw new WalletNotFoundException();
+        }
+        if ($wallet->hasBalance()) {
+            throw new CouldNotDeleteWalletWithBalanceException();
+        }
+        if ($wallet->isDefault()) {
+            throw new CouldNotDeleteWalletWithBalanceException('could not delete default wallet');
+        }
+
         try {
             return $this->repository->delete($id);
         }
-        catch (Exception $exception) {
+        catch (Exception $e) {
+            XLog::exception($e);
             throw new DeleteResourceFailedException();
         }
     }
